# openapi.yaml
openapi: 3.0.0
info:
  title: CSC309 PP1
  description: >-
    This collection provides RESTful JSON endpoints for managing accounts,
    hotels, flights, bookings, and notifications. All protected endpoints
    require a valid JWT (set in the Authorization header as Bearer {{token}}).
    The APIs support user registration/login, hotel and room management, flight
    search through the AFS system, booking operations (create, checkout,
    cancel), and notifications. Additionally, there are additional endpoints for
    the locally hosted lists of cities and airports for frontend use (to come
    later in PP2).


    Below you will find all the user stories we implemented for PP1.


    ### **Accounts**


    - **U1: As a user, I want to sign up, log in, log out, and edit my profile.
    Profile information includes first and last name, email, profile picture,
    and phone number. Authentication should be handled with a proper JWT
    setup.**
        

    ---


    ### **Flights Search**


    - **U2: As a visitor, I want to search for flights by specifying a source,
    destination, and date(s). Source and destination could be either a city or
    an airport. I want to search for one-way or round-trip flights.**
        
    - **U3: As a visitor, I want to view flight details, including
    departure/arrival times, duration, and layovers.**
        

    _(The stories about auto-complete and date picker dropdowns are UI concerns
    and will be implemented in Part 2.)_


...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /api/suggestions:
    post:
      tags:
        - Booking
      summary: Get Suggestions
      description: >-
        Returns booking suggestions for flights and hotels based on input
        criteria.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                hotelId: 18
                destination: Zurich
                departureCity: Toronto
                suggestedDate: '2025-04-15'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Suggestions retrieved successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  hotels:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        starRating:
                          type: integer
                        address:
                          type: string
                        images:
                          type: array
                          items:
                            type: string
                  searchUrl:
                    type: string
                  flights:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        airline:
                          type: object
                          properties:
                            code:
                              type: string
                            name:
                              type: string
                        departureTime:
                          type: string
                          format: date-time
                        price:
                          type: integer
              example:
                hotels:
                  - id: 16
                    name: "Swiss Inn"
                    starRating: 5
                    address: "123 Alpine Rd"
                    images:
                      - "http://example.com/img1.png"
                      - "http://example.com/img2.png"
                  - id: 18
                    name: "Swiss Luxury"
                    starRating: 5
                    address: "456 Bahnhofstrasse"
                    images:
                      - "http://example.com/img1.png"
                searchUrl: "/search/flights?destination=Zurich&keepOrder=true"
                flights:
                  - id: "e29b88fa-551d-4bb9-b7dd-99333c860efe"
                    airline:
                      code: "AC"
                      name: "Air Canada"
                    departureTime: "2025-04-15T00:55:00.000Z"
                    price: 1563
                  - id: "66d9a5fd-7d2d-4491-bbc6-bdf0a15fb7a5"
                    airline:
                      code: "SK"
                      name: "Scandinavian Airlines"
                    departureTime: "2025-04-15T17:15:00.000Z"
                    price: 1672
                  - id: "7ae43949-9928-4466-b6ef-48959f16a43a"
                    airline:
                      code: "AC"
                      name: "Air Canada"
                    departureTime: "2025-04-15T06:20:00.000Z"
                    price: 537
                  - id: "447fef24-df4b-4733-a5f9-1c511b0228d3"
                    airline:
                      code: "B6"
                      name: "JetBlue"
                    departureTime: "2025-04-16T07:05:00.000Z"
                    price: 1789
...